move new steerable rotate tests out of test_rotation
