+++ tests/test_rotation.py
@@ -1,13 +1,10 @@
 import logging
 
 import numpy as np
-import PIL.Image as PILImage
 import pytest
 from scipy.spatial.transform import Rotation as sp_rot
 
-from aspire.basis import FBBasis2D, FFBBasis2D, FLEBasis2D, FPSWFBasis2D, PSWFBasis2D
-from aspire.image import Image
-from aspire.utils import Rotation, gaussian_2d, utest_tolerance
+from aspire.utils import Rotation, utest_tolerance
 
 logger = logging.getLogger(__name__)
 
@@ -19,21 +16,9 @@ SEED = 0
 
 DTYPES = [
     np.float32,
-    pytest.param(np.float64, marks=pytest.mark.expensive),
+    np.float64,
 ]
 
-BASES = [
-    FFBBasis2D,
-    FBBasis2D,
-    FLEBasis2D,
-    PSWFBasis2D,
-    FPSWFBasis2D,
-]
-
-IMG_SIZES = [
-    32,
-    pytest.param(31, marks=pytest.mark.expensive),
-]
 
 # Fixtures
 
@@ -46,19 +31,6 @@ def dtype(request):
 @pytest.fixture(scope="module")
 def rot_obj(dtype):
     return Rotation.generate_random_rotations(NUM_ROTS, seed=SEED, dtype=dtype)
-
-
-@pytest.fixture(params=IMG_SIZES, ids=lambda x: f"img_size={x}", scope="module")
-def img_size(request):
-    return request.param
-
-
-@pytest.fixture(params=BASES, ids=lambda x: f"basis={x}", scope="module")
-def basis(request, img_size, dtype):
-    cls = request.param
-    # Setup a Basis
-    basis = cls(img_size, dtype=dtype)
-    return basis
 
 
 # Rotation Class Tests
@@ -207,33 +179,3 @@ def test_mean_angular_distance(dtype):
     mean_ang_dist = Rotation.mean_angular_distance(rots_z, rots_id)
 
     assert np.allclose(mean_ang_dist, np.pi / 4)
-
-
-# Basis Rotations
-
-
-def test_basis_rotation_2d(basis):
-    """
-    Test steerable basis rotation performs similar operation to PIL real space image rotation.
-
-    Checks both orientation and rough values.
-    """
-    # Set a rotation amount
-    rot_radians = np.pi / 6
-
-    # Create an Image containing a smooth blob.
-    L = basis.nres
-    img = Image(gaussian_2d(L, mu=(L // 4, 0), dtype=basis.dtype))
-
-    # Rotate with ASPIRE Steerable Basis, returning to real space.
-    rot_img = basis.expand(img).rotate(rot_radians).evaluate()
-
-    # Rotate image with PIL, returning to Numpy array.
-    pil_rot_img = np.asarray(
-        PILImage.fromarray(img.asnumpy()[0]).rotate(
-            rot_radians * 180 / np.pi, resample=PILImage.BILINEAR
-        )
-    )
-
-    # Rough compare arrays.
-    np.testing.assert_allclose(rot_img.asnumpy()[0], pil_rot_img, atol=0.25)
